Cambio en el metodo multiplicador()
La variable de salida no puede ser una variable tipo entero, ya que hay salidas decimales.

+++ src/tipos.h
@@ -39,7 +39,7 @@ class Tipo{;
         Tipos getSecondary();
         string getPrimaryName();
         string getSecondaryName();
-        uint multiplicador(Tipos tipo);
+        float multiplicador(Tipos tipo);
 }
 
 #endif // TIPOS_H
